requirements.txt
python-telegram-bot==12.8
requests==2.24.0

boto3~=1.15.1
certifi~=2020.6.20
chardet~=3.0.4
tornado~=6.0.4
cryptography~=3.1
urllib3~=1.25.10
idna~=2.10
decorator~=4.4.2
six~=1.15.0
pycparser~=2.20